split notifications
