+++ shared/constants/notifications.js
@@ -1,18 +1,6 @@
 // @flow
 import * as I from 'immutable'
-import type {Tab} from './tabs'
-
-export type NotificationKeys = 'kbfsUploading'
-export type BadgeType = 'regular' | 'update' | 'badged' | 'uploading'
-
-type _State = {
-  desktopAppBadgeCount: number,
-  keyState: I.Map<NotificationKeys, boolean>,
-  mobileAppBadgeCount: number,
-  navBadges: I.Map<Tab, number>,
-  widgetBadge: BadgeType,
-}
-export type State = I.RecordOf<_State>
+import type {_State} from './types/notifications'
 
 export const makeState: I.RecordFactory<_State> = I.Record({
   desktopAppBadgeCount: 0,
