+++ shared/constants/reducer.js
@@ -9,7 +9,7 @@ import type {State as Engine} from '../constants/types/engine'
 import type {State as Favorite} from '../constants/types/favorite'
 import type {State as Gregor} from '../constants/types/gregor'
 import type {State as Login} from '../constants/types/login'
-import type {State as Notification} from '../constants/notifications'
+import type {State as Notification} from '../constants/types/notifications'
 import type {State as Pgp} from '../constants/pgp'
 import type {State as Pinentry} from '../constants/pinentry'
 import type {State as Profile} from '../constants/profile'
